+++ src/components/Pages/Home.js
@@ -1,9 +1,19 @@
 import React from "react";
+import "./Home.css"
+import ImageSlider from "./Images/ImageSlider";
+import { SliderData } from './Images/SliderData';
+import HomePageImages from "./Images/HomePageImages";
 
-function Home(){
+
+function Home() {
     return(
-        <h1>Test</h1>
+    <div>
+        <h1> FEATURED </h1>
+        <ImageSlider slides={SliderData} />
+        <h1>WHAT'S ON</h1>
+        <HomePageImages />
+    </div>
     )
 };
 
-export default Home
+export default Home;
